Auth

+++ frontend/src/components/WorkoutDetails.js
@@ -1,4 +1,6 @@
 import { useWorkoutContext } from "../hooks/useWorkoutContext"
+import { useAuthContext } from "../hooks/useAuthContext"
+
 
 //date-fns
 import { formatDistanceToNow } from 'date-fns'
@@ -7,9 +9,22 @@ const WorkoutDetails = ({workout}) => {
 
     const { dispatch } = useWorkoutContext()
 
+    const { user } = useAuthContext()
+
+
+
     const handleClick = async() => {
+
+        if(!user) {
+            return
+        }
+        
         const response = await fetch('/api/workouts/' + workout._id, {
             method: 'DELETE',
+            headers: {
+            'Authorization': `Bearer ${user.token}`
+            }
+
         })
         const json = await response.json()
